Create separate classes for data parsing -Delete old code
