+++ pymt/models/oasth/Line.py
@@ -1,0 +1,4 @@
+class Line(object):
+    def __init__(self, d):
+        self.__dict__ = d
+        self.stops = []
