+++ pymt/models/oasth/Stop.py
@@ -1,0 +1,3 @@
+class Stop(object):
+    def __init__(self, d):
+        self.__dict__ = d
