+++ pymt/models/oasth/__init__.py
@@ -1,0 +1,5 @@
+from . import Line, Bus, Stop
+
+Stop = Stop.Stop
+Line = Line.Line
+Bus = Bus.Bus
