+++ pymt/mongo_functions.py
@@ -1,5 +1,6 @@
 from pymongo import MongoClient
 from pymt import config
+import pymt.models.oasth as model
 
 c = config['mongodb']
 
@@ -13,31 +14,20 @@ stops_db = db[c['stops_db']]
 route_stops_db = db[c['route_stops_db']]
 
 
-class CustomStop(object):
-    def __init__(self, d):
-        self.__dict__ = d
-
-
-class CustomLine(object):
-    def __init__(self, d):
-        self.__dict__ = d
-        self.stops = []
-
-
 # TODO: fix encoding in route names
 def get_route_stops(name='31'):
     line = lines_db.find_one(dict(name=name))
     route = line.get('days')[0]
     # stops = route_stops_db.find(dict(route_id=route))
     stop_ids = [s.get('stop_id') for s in route_stops_db.find(dict(route_id=route))]
-    stops = [CustomStop(stops_db.find_one(dict(_id=s))) for s in stop_ids]
+    stops = [model.Stop(stops_db.find_one(dict(_id=s))) for s in stop_ids]
     return stops
 
 
 def get_line_by_name(name):
-    line = CustomLine(lines_db.find_one(dict(name=name)))
+    line = model.Line(lines_db.find_one(dict(name=name)))
     return line
 
 
 def get_all_lines():
-    return [CustomLine(l) for l in lines_db.find()]
+    return [model.Line(l) for l in lines_db.find()]
